Create EmailAppScript
This is the code to automated the sending of emails using Google App Script. - https://script.google.com/ - The email addresses have to be stored in the a google sheet under the same account as the email account and it has to have an index so it moves to the next  row after sending an email to that cells address. Should you have any questions, please feel free to contact me.